@@ -1,9 +1,9 @@
 import { render, screen } from '@testing-library/react';
+import React from "react"
 import App from './App';
-import React from "react";
-import '@testing-library/jest-dom'
 
 test('renders learn react link', () => {
   render(<App />);
-  expect(screen.getByRole('add')).not.toBeDisabled()
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
 });
